third_party/WebKit/LayoutTests/inspector/sources/debugger-step/debugger-step-into-async1-expected.txt

Tests debugger StepIntoAsync action (part 1).

Set timer for test function.
Executing StepOver...
Executing StepOver...
Executing StepOver...
Call stack:
    0) testFunction (debugger-step-into-async1.html:17)

Executing StepIntoAsync...
Call stack:
    0) callback1 (debugger-step-into-async1.html:23)
    [setTimeout]
    0) testFunction (debugger-step-into-async1.html:17)

Executing Resume...
Executing StepOver...
Executing StepIntoAsync...
Call stack:
    0) callback2 (debugger-step-into-async1.html:34)
    [setTimeout]
    0) callback1 (debugger-step-into-async1.html:29)
    [setTimeout]
    0) testFunction (debugger-step-into-async1.html:17)

Executing Resume...
Executing StepOver...
Executing StepIntoAsync...
Call stack:
    0) callback3 (debugger-step-into-async1.html:46)
    [setTimeout]
  * 0) Framework.schedule (framework.js:45)
  * 1) Framework_willSchedule (framework.js:51)
    [setTimeout]
  * 0) Framework.schedule (framework.js:45)
  * 1) Framework_willSchedule (framework.js:51)
    [setTimeout]
  * 0) Framework.schedule (framework.js:45)
  * 1) Framework_willSchedule (framework.js:51)
    [setTimeout]
  * 0) Framework.schedule (framework.js:45)
    1) callback2 (debugger-step-into-async1.html:40)

Executing Resume...
Executing StepOver...
Executing StepIntoAsync...
Call stack:
    0) callback4 (debugger-step-into-async1.html:56)
  * 1) Framework.safeRun (framework.js:8)
  * 2) Framework.safeRun (framework.js:13)
  * 3) Framework.safeRun (framework.js:10)
  * 4) Framework_bound (framework.js:105)
    [setTimeout]
  * 0) Framework.schedule (framework.js:45)
    1) callback3 (debugger-step-into-async1.html:50)
    [setTimeout]
  * 0) Framework.schedule (framework.js:45)
  * 1) Framework_willSchedule (framework.js:51)
    [setTimeout]
  * 0) Framework.schedule (framework.js:45)
  * 1) Framework_willSchedule (framework.js:51)
    [setTimeout]
  * 0) Framework.schedule (framework.js:45)
  * 1) Framework_willSchedule (framework.js:51)

Executing Resume...
Executing StepOver...
Executing StepIntoAsync...
Call stack:
    0) callback5 (debugger-step-into-async1.html:64)
  * 1) Framework_eventListener (framework.js:86)
  * 2) Framework.createButtonWithEventListenersAndClick (framework.js:173)
  * 3) Framework_bound (framework.js:105)
    [setTimeout]
  * 0) Framework.schedule (framework.js:45)
    1) callback4 (debugger-step-into-async1.html:58)
  * 2) Framework.safeRun (framework.js:8)
  * 3) Framework.safeRun (framework.js:13)
  * 4) Framework.safeRun (framework.js:10)
  * 5) Framework_bound (framework.js:105)
    [setTimeout]
  * 0) Framework.schedule (framework.js:45)
    1) callback3 (debugger-step-into-async1.html:50)
    [setTimeout]
  * 0) Framework.schedule (framework.js:45)
  * 1) Framework_willSchedule (framework.js:51)
    [setTimeout]
  * 0) Framework.schedule (framework.js:45)
  * 1) Framework_willSchedule (framework.js:51)

Executing Resume...
Executing StepOver...
Executing StepIntoAsync...
Call stack:
    0) callback6 (debugger-step-into-async1.html:80)
    [Promise.resolve]
    0) callback5 (debugger-step-into-async1.html:75)
  * 1) Framework_eventListener (framework.js:86)
  * 2) Framework.createButtonWithEventListenersAndClick (framework.js:173)
  * 3) Framework_bound (framework.js:105)
    [setTimeout]
  * 0) Framework.schedule (framework.js:45)
    1) callback4 (debugger-step-into-async1.html:58)
  * 2) Framework.safeRun (framework.js:8)
  * 3) Framework.safeRun (framework.js:13)
  * 4) Framework.safeRun (framework.js:10)
  * 5) Framework_bound (framework.js:105)
    [setTimeout]
  * 0) Framework.schedule (framework.js:45)
    1) callback3 (debugger-step-into-async1.html:50)
    [setTimeout]
  * 0) Framework.schedule (framework.js:45)
  * 1) Framework_willSchedule (framework.js:51)

Executing Resume...
Executing StepOver...
Executing StepIntoAsync...
Call stack:
    0) callback7 (debugger-step-into-async1.html:96)
    [Promise.resolve]
    0) callback6 (debugger-step-into-async1.html:85)
    [Promise.resolve]
    0) callback5 (debugger-step-into-async1.html:75)
  * 1) Framework_eventListener (framework.js:86)
  * 2) Framework.createButtonWithEventListenersAndClick (framework.js:173)
  * 3) Framework_bound (framework.js:105)
    [setTimeout]
  * 0) Framework.schedule (framework.js:45)
    1) callback4 (debugger-step-into-async1.html:58)
  * 2) Framework.safeRun (framework.js:8)
  * 3) Framework.safeRun (framework.js:13)
  * 4) Framework.safeRun (framework.js:10)
  * 5) Framework_bound (framework.js:105)
    [setTimeout]
  * 0) Framework.schedule (framework.js:45)
    1) callback3 (debugger-step-into-async1.html:50)
